http: Allow the msgpack content-type, reject the request post-auth.
This enables the msgpack content-type messages to make it down to the
processor and then finally rejected with a 501 before attempting to
process.

A few changes were made to support easier testing, such as the move of
the individual HandlerFunc funcs to concrete functions for use directly
in a test server as well as the introduction of the nullForwarder so
forwards are no-ops.