Adding support for logging the auth user provided
Setting up for rolling logs.herokai.com creds, we're adding the user
used to authenticate to log-iss to each log line. This should help us
determine when it is safe to remove the old cred. I'm putting this
behind a feature flag so we can toggle it off once the cred roll is
complete should we so choose.