Make metadata id customizable.
According to RFC 5424 6.3.2 (https://tools.ietf.org/html/rfc5424#section-6.3.2), structured data SD-ID must be in the format "name@<private-enterprise-number>", unless the structured data is one of the reserved fields. The data we are adding is *not* a reserved field, so we need to make it customizable. We'll use something like "metadata@233507" for our value for this (see https://www.iana.org/assignments/enterprise-numbers/enterprise-numbers, look for 23507).