getMetadata refactor
This was part of ffz/Optimize. I'm very sure this wasn't part of the
problem.

This dedents most of the logic & removes (what is in my opinion) the
unnecessary bool return (it's a guard on the length of the bytes).

Dedenting the majority of the logic in a function a go best practice.